refactor(map): extract geo / meeting / visited-stops hooks (3/7)
Step 3 of the PulseMapView strangler decomposition. Pure move — no
behaviour change. Effect dependencies, memo composition, and the
visited-stops marker-key construction are preserved byte-equivalent.

services/geofenceService.ts
  - New read API: getTodayVisitedContactStops() returns Set<string> of
    marker keys (`${contactId}-home` / `${contactId}-work`) for today's
    geofence enter events with entity_type='contact'. Conservative on
    role: marks both home and work when the row carries no role hint,
    matching the prior in-line behaviour.

components/map/hooks/useGeoRelevanceSignals.ts
  - GeoSignals shape, contactAttendsEvent + lensIncludesContact
    predicates, and the useGeoRelevanceSignals hook. The hook owns
    fetchedEvents / fetchedThreads state and the dataService self-fetch
    effect; the memo composes the bundle from props-or-self-fetch
    exactly as before.

components/map/hooks/useMeetingMarkers.ts
  - Carries the MeetingMarkerData type and the lazy geocode effect with
    its geocodedEventsRef cache. Same eslint-disable on the deps array
    (the prior in-line note about not loop-firing on meetingMarkers
    still applies after the move).

components/map/hooks/useVisitedStops.ts
  - Calls geofenceService.getTodayVisitedContactStops on lens change.
    Returns the Set the AI proposal effect already consumes.

PulseMapView.tsx
  - Replaces three inline blocks (~145 lines) with three hook calls
    plus a single multi-line orientation comment.
  - Imports lensIncludesContact from the new module — the visibleMarkers
    filter and the AI proposal effect both still call it.
  - Drops Thread, dataService, supabase, geocodeAddress, and WEEK_MS
    imports (all only used inside the extracted blocks).
